+++ firmware/src/address_bus.h
@@ -1,3 +1,3 @@
 void address_bus_clock();
 void address_bus_latch();
-void address_bus_set(unsigned int value);
+void address_bus_set(uint16_t value);
